feat: Rename the CLI tool from migrate to db (MPT-12683)
