+++ tools/db/src/db/__main__.py
@@ -1,0 +1,3 @@
+from db.cli import app
+
+app()
