--- tools/migrate/src/migrate/__main__.py
@@ -1,3 +1,0 @@
-from migrate.cli import app
-
-app()
