--- tools/migrate/src/migrate/migrators/__init__.py
@@ -1,6 +1,0 @@
-from migrate.migrators.alembic import AlembicMigrator
-from migrate.migrators.mongo import MongoMigrator
-from migrate.migrators.clickhouse import ClickhouseMigrator
-from migrate.migrators.base import BaseMigrator
-
-__all__ = ["AlembicMigrator", "MongoMigrator", "ClickhouseMigrator", "BaseMigrator"]
